Initial commit of simple CC310 support and example.
This provides basic support for using CC310 to accelrate SHA-1 and
AES CBC 128 operations. This must be done in secure region and is
therefore, implemented in the SPM, which provides non-secure callable
functions that can be used by a non-secure application.

An accompanying application example demonstrates the use and also
does the same operations with mbed TLS SW libra for comparison.